Address code reviews
- Changes return type of futex() to InterpResult<'tcx> and removes every ok() calls
- Add documentation and link to `_umtx_op` man pages
- Move comments to parameters and remove pointless assignments
- Replace `OpTy`s usages to `Pointer`s
- Run `deref_operand` on `read_scalar` output in Futex wait operation
- Replace `to_u64` to `to_machine_usize`
