+++ src/shims/unix/freebsd/mod.rs
@@ -1,3 +1,3 @@
 pub mod dlsym;
 pub mod foreign_items;
-pub mod futex;
+pub mod sync;
